Validators... Read the description
We only need to check if the zone is a validator zone -> Yes, count Player.Validator + 1 (No, do nothing).

And we hit the checker zone, we need to check if it's a checker zone and check his Validator effect value. This must be the amount of validators of this linear/checkpoint map or of the stage.

Why this way? We reset the validator hits in every start/end/stage zone touch ( This must be documented in our wiki/faq ).